refactor: dedupe shared payload helpers across remote executors
The Kubernetes and GCP Batch executors carried byte-for-byte copies of
four helpers (_encode_payload, _payload_requires_fuse,
_parse_worker_output, and _generate_job_id/_generate_job_name). Move the
shared logic into ginkgo.remote._executor_common and consolidate the job
naming function under a single name — the two implementations were
already producing identical names compatible with both Kubernetes and
GCP Batch constraints.

The kubernetes.py module keeps a thin _parse_worker_output wrapper so
its fallback error message still says "pod logs" instead of "job logs".

Addresses findings 3-6 of issue #53.
